refactor module_store from move_module
