+++ frameworks/moveos-stdlib/doc/README.md
@@ -32,6 +32,7 @@ This is the reference documentation of the MoveOS standard library.
 -  [`0x2::hash`](hash.md#0x2_hash)
 -  [`0x2::hex`](hex.md#0x2_hex)
 -  [`0x2::json`](json.md#0x2_json)
+-  [`0x2::module_store`](module_store.md#0x2_module_store)
 -  [`0x2::move_module`](move_module.md#0x2_move_module)
 -  [`0x2::object`](object.md#0x2_object)
 -  [`0x2::rlp`](rlp.md#0x2_rlp)
