+++ frameworks/moveos-stdlib/doc/genesis.md
@@ -11,7 +11,7 @@
 <pre><code><b>use</b> <a href="">0x1::option</a>;
 <b>use</b> <a href="features.md#0x2_features">0x2::features</a>;
 <b>use</b> <a href="gas_schedule.md#0x2_gas_schedule">0x2::gas_schedule</a>;
-<b>use</b> <a href="move_module.md#0x2_move_module">0x2::move_module</a>;
+<b>use</b> <a href="module_store.md#0x2_module_store">0x2::module_store</a>;
 <b>use</b> <a href="tx_context.md#0x2_tx_context">0x2::tx_context</a>;
 </code></pre>
