+++ frameworks/moveos-stdlib/sources/genesis.move
@@ -3,7 +3,7 @@
 
 module moveos_std::genesis {
     use std::option;
-    use moveos_std::move_module;
+    use moveos_std::module_store;
     use moveos_std::features;
     use moveos_std::tx_context;
     use moveos_std::gas_schedule::{Self, GasScheduleConfig};
@@ -11,7 +11,7 @@ module moveos_std::genesis {
     const ErrorGenesisInit: u64 = 1;
 
     fun init(){
-        move_module::init_module_store();
+        module_store::init_module_store();
         features::init_feature_store();
         let gas_config_option = tx_context::get_attribute<GasScheduleConfig>();
         assert!(option::is_some(&gas_config_option), ErrorGenesisInit);
