+++ frameworks/rooch-framework/doc/onchain_config.md
@@ -19,7 +19,7 @@
 
 
 <pre><code><b>use</b> <a href="">0x2::features</a>;
-<b>use</b> <a href="">0x2::move_module</a>;
+<b>use</b> <a href="">0x2::module_store</a>;
 <b>use</b> <a href="">0x2::object</a>;
 <b>use</b> <a href="">0x2::signer</a>;
 <b>use</b> <a href="chain_id.md#0x3_chain_id">0x3::chain_id</a>;
