+++ frameworks/rooch-framework/doc/upgrade.md
@@ -12,7 +12,7 @@
 
 <pre><code><b>use</b> <a href="">0x2::account</a>;
 <b>use</b> <a href="">0x2::event</a>;
-<b>use</b> <a href="">0x2::move_module</a>;
+<b>use</b> <a href="">0x2::module_store</a>;
 <b>use</b> <a href="">0x2::signer</a>;
 <b>use</b> <a href="onchain_config.md#0x3_onchain_config">0x3::onchain_config</a>;
 </code></pre>
