+++ frameworks/rooch-framework/sources/onchain_config.move
@@ -6,7 +6,7 @@ module rooch_framework::onchain_config {
     use std::vector;
     use moveos_std::object;
     use moveos_std::features;
-    use moveos_std::move_module;
+    use moveos_std::module_store;
     use moveos_std::signer;
     use rooch_framework::chain_id;
 
@@ -63,16 +63,16 @@ module rooch_framework::onchain_config {
     public entry fun add_to_publishing_allowlist(account: &signer, publisher: address) {
         ensure_sequencer(account);
         let system_account = signer::module_signer<OnchainConfig>();
-        let allowlist = move_module::borrow_mut_allowlist();
-        move_module::add_to_allowlist(allowlist, &system_account, publisher);
+        let allowlist = module_store::borrow_mut_allowlist();
+        module_store::add_to_allowlist(allowlist, &system_account, publisher);
     }
 
     /// Remove `publisher` from publishing allowlist.
     public entry fun remove_from_publishing_allowlist(account: &signer, publisher: address) {
         ensure_sequencer(account);
         let system_account = signer::module_signer<OnchainConfig>();
-        let allowlist = move_module::borrow_mut_allowlist();
-        move_module::remove_from_allowlist(allowlist, &system_account, publisher);
+        let allowlist = module_store::borrow_mut_allowlist();
+        module_store::remove_from_allowlist(allowlist, &system_account, publisher);
     }
     /****** End of API for update module publishing allowlist. ******/
