+++ moveos/moveos-types/src/moveos_std/mod.rs
@@ -9,6 +9,7 @@ pub mod copyable_any;
 pub mod display;
 pub mod event;
 pub mod gas_schedule;
+pub mod module_store;
 pub mod module_upgrade_flag;
 pub mod move_module;
 pub mod object;
