+++ moveos/moveos-types/src/state_resolver.rs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::moveos_std::account::Account;
-use crate::moveos_std::move_module::ModuleStore;
+use crate::moveos_std::module_store::ModuleStore;
 use crate::moveos_std::object::{ObjectID, RawObject, RootObjectEntity};
 use crate::state::{AnnotatedKeyState, KeyState};
 use crate::{
